refactor : stacked architecture
