+++ lib/ui/widgets/smart_widgets/URL BottomSheet/url_bottomsheet_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:rutorrentflutter/AppTheme/AppStateNotifier.dart';
 import 'package:rutorrentflutter/enums/enums.dart';
-import 'package:rutorrentflutter/ui/widgets/dumb_widgets/data_input.dart';
+import 'package:rutorrentflutter/ui/widgets/dumb_widgets/data_input_widget.dart';
 import 'package:rutorrentflutter/ui/widgets/smart_widgets/URL%20BottomSheet/url_bottomsheet_viewmodel.dart';
 import 'package:stacked/stacked.dart';
